third_party/WebKit/LayoutTests/fast/forms/input-inputmode-expected.txt

Tests the behavior of .inputMode of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.inputMode is ""
PASS input.inputMode is "0"
PASS input.inputMode is "abc"
PASS input.getAttribute("inputmode") is "foo"
PASS input.getAttribute("inputmode") is ""
Setting null to inputMode:
PASS input.inputMode is "null"
PASS input.getAttribute("inputmode") is "null"
PASS input.inputMode is "null"
Setting undefined to inputMode:
PASS input.inputMode is "undefined"
PASS input.getAttribute("inputmode") is "undefined"
PASS input.inputMode is "undefined"
Setting non-string to inputMode:
PASS input.inputMode is "256"
PASS input.getAttribute("inputmode") is "256"
PASS input.inputMode is "256"
PASS successfullyParsed is true

TEST COMPLETE
